Add rowKey function to Tuple/PrivateTuple typeclass
For token aware routing, we need to be able to turn a tuple into a
binary value given the primary key indices. Generate instances for
tuples that turn a tuple into a vector of values, and then use constant
time lookup for the values. Building the key will fail if the primary
key indices are out of bounds of the vector.