Add a utility function for the Cassandra Murmur3 hash
In order to correctly route tokens for a token-aware policy, it is necessary to
correctly hash any private keys. Cassandra unfortunately doesn't match the
standard murmur3 hash function, so an off the shelf library for that isn't an
option. Import the library from https://github.com/russss/murmur3-cassandra,
complete with its license, and export a pure wrapper from ByteString to Int64.