Separate putValue and putValueLength
Make two versions of `putValue`, one that prefixes with the length, and
one that doesn't. The one that prefixes with the length calls `putValue`
in non-trivial cases.

This makes building correct row keys easier. Row keys are not prefixed
with the initial length.

Also clean up IP address handling. IP addresses are all serialized in
big endian order in the native protocol. Only use host encoding in the
one place we need it. IPv4 addresses are stored as big-endian order
regardless of the host endianness. See the documentation in the network
package. Everywhere else, we can use a defined endianness. Also,
wherever possible, we use an explicit endianness instead of relying on
the fact that the default for cereal is big endian.