Use persist to avoid re-serialization
In multiple places, in order to serialize a value prepended by a length,
we would serialize the value completely to a bytestring, then serialize
the length, and then serialize the serialized bytestring. For a nested
data structure, this could lead to multiple layers of copying.

Take advantage of a new feature in the `persist` library starting with
the 1.0 release. `persist` has the ability to allocate space for a
length and then fill in the length afterwords. This means that we run
the serializer exactly once, without copying up the nesting tree.